stock_analysis/naptha_modules/financial_agent/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "financial_agent"
version = "v0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
crewai = "^0.100.0"
crewai_tools = ">=0.5.0"
pandas = ">=2.2.3"
pydantic = "*"
sec_api = "*"
embedchain = ">=0.1.113,<0.2.0"
requests = "*"
html2text = "*"
python-dotenv = "*"

[tool.poetry.dependencies.naptha-sdk]
git = "https://github.com/js-ts/naptha-sdk.git"
branch = "main"
